requirements.txt
flask-restplus
Flask-SQLAlchemy
pymysql
flask-jwt-extended